// src/app/models/estadosPedidosEnum.ts
export enum EstadosPedidos {
  PENDIENTE = 'Pendiente',
  EN_AUTORIZACION = 'En Autorizacion',
  EN_PREPARACION = 'En Preparacion',
  EN_EMPAQUETADO = 'En Empaquetado',
  EN_DESPACHO = 'En Despacho',
  EN_DISTRIBUCION = 'En Distribucion',
  EN_FACTURACION = 'En Facturacion',
  COMPLETADO = 'Completado',
  CANCELADO = 'Cancelado'
}